Lagt til et rest-endepunkt som kan brukes for å opprette test-eventene som trengs til ytelsestesten. Bruker egne topics til denne testen.
PB-138. Ytelsestesting av DEA
