+++ src/main/kotlin/no/nav/personbruker/dittnav/eventtestproducer/config/Kafka.kt
@@ -17,10 +17,10 @@ object Kafka {
 
     val log = LoggerFactory.getLogger(Kafka::class.java)
 
-    val doneTopicName = "aapen-brukernotifikasjon-done-v1"
-    val beskjedTopicName = "aapen-brukernotifikasjon-nyBeskjed-v1"
+    val doneTopicName = "aapen-brukernotifikasjon-done-v1-ytelsestesting"
+    val beskjedTopicName = "aapen-brukernotifikasjon-nyBeskjed-v1-ytelsestesting"
     val innboksTopicName = "aapen-brukernotifikasjon-nyInnboks-v1"
-    val oppgaveTopicName = "aapen-brukernotifikasjon-nyOppgave-v1"
+    val oppgaveTopicName = "aapen-brukernotifikasjon-nyOppgave-v1-ytelsestesting"
 
     private fun credentialProps(env: Environment): Properties {
         return Properties().apply {
