+++ src/main/kotlin/no/nav/personbruker/dittnav/eventtestproducer/config/bootstrap.kt
@@ -22,6 +22,7 @@ import no.nav.personbruker.dittnav.eventtestproducer.common.healthApi
 import no.nav.personbruker.dittnav.eventtestproducer.done.doneApi
 import no.nav.personbruker.dittnav.eventtestproducer.innboks.innboksApi
 import no.nav.personbruker.dittnav.eventtestproducer.oppgave.oppgaveApi
+import no.nav.personbruker.dittnav.eventtestproducer.ytelsestesting.ytelsestestApi
 import no.nav.security.token.support.ktor.tokenValidationSupport
 
 @KtorExperimentalAPI
@@ -55,6 +56,7 @@ fun Application.mainModule(appContext: ApplicationContext = ApplicationContext()
             beskjedApi(appContext.beskjedProducer)
             innboksApi(appContext.innboksProducer)
             doneApi(appContext.doneEventService)
+            ytelsestestApi(appContext.testDataService)
         }
     }
